Close #25; comments on Salvo and Simon's papers distributed through the text.
Where we discuss the "power" of chieff > 0, reference to Salvo's paper, with 10--20 sources required to get 3\sigma opposed to our 10 for >> 5 sigma.

Where we discuss difficulty of determining spin magnitudes, refer to recent studies in support.

Previous para devoted to Salvo, Simon, and spin magnitudes now only cautions that if multiple populations present to significant degree, may required 10s to 100s of detections, citing Salvo and Simon, depending on magnitudes.